pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "The Difference Calculator. Shows the difference between two data files"
authors = [{name = "Vitaliy Berezhnoy", email = "ipbvv@mail.ru"}]
maintainers = [{name = "Vitalii Vorobev"}]
repository = "https://github.com/Vitaliy-Berezhnoy/python-project-50.git"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = 'gendiff.scripts.gendiff:main'
